Create Base Model
We will create the base model from the MobileNet V2 model developed at Google, which is pre-trained on the ImageNet dataset. This dataset includes 1.4 million images and 1,000 classes, encompassing a wide range of categories such as jackfruit and syringe. This extensive base of knowledge will assist us in classifying whether a person is wearing a mask or not using our specific dataset.

First, we need to decide which layer of MobileNet V2 to use for extracting high-level features. To adapt the model's classifications for a new task, we should utilize the features from the last layer before the classification layers. Typically, in many image models, this is the output of the final convolution layer before the flatten layer, often referred to as the "bottleneck layer" in some texts. Since many machine learning models have inputs at the bottom and outputs at the top, we should ignore the top few classification layers. Fortunately, TensorFlow offers a shortcut for this by using the parameter include_top=False. By passing this parameter, we instantiate a MobileNet V2 model pre-loaded with weights trained on ImageNet, excluding the top classification layers, which is ideal for feature extraction.